+++ packages/webview_flutter/webview_flutter_wkwebview/ios/Classes/FWFDataConverters.m
@@ -211,6 +211,8 @@ FWFNSKeyValueChangeKeyEnumData *FWFNSKeyValueChangeKeyEnumDataFromNativeNSKeyVal
         makeWithValue:FWFNSKeyValueChangeKeyEnumNotificationIsPrior];
   } else if ([key isEqualToString:NSKeyValueChangeOldKey]) {
     return [FWFNSKeyValueChangeKeyEnumData makeWithValue:FWFNSKeyValueChangeKeyEnumOldValue];
+  } else {
+    return [FWFNSKeyValueChangeKeyEnumData makeWithValue:FWFNSKeyValueChangeKeyEnumUnknown];
   }
 
   return nil;
@@ -235,6 +237,8 @@ FWFWKNavigationType FWFWKNavigationTypeFromNativeWKNavigationType(WKNavigationTy
     case WKNavigationTypeOther:
       return FWFWKNavigationTypeOther;
   }
+
+  return FWFWKNavigationTypeUnknown;
 }
 
 FWFWKSecurityOriginData *FWFWKSecurityOriginDataFromNativeWKSecurityOrigin(