+++ packages/webview_flutter/webview_flutter_wkwebview/lib/src/common/web_kit.g.dart
@@ -40,6 +40,7 @@ enum NSKeyValueChangeKeyEnum {
   newValue,
   notificationIsPrior,
   oldValue,
+  unknown,
 }
 
 /// Mirror of WKUserScriptInjectionTime.
@@ -136,6 +137,12 @@ enum WKNavigationType {
   ///
   /// See https://developer.apple.com/documentation/webkit/wknavigationtype/wknavigationtypeother?language=objc.
   other,
+
+  /// An unknown navigation type.
+  ///
+  /// This does not represent an actual value provided by the platform and only
+  /// indicates a value was provided that isn't currently supported.
+  unknown,
 }
 
 /// Possible permission decisions for device resource access.
@@ -180,7 +187,7 @@ enum WKMediaCaptureType {
   /// An unknown media device.
   ///
   /// This does not represent an actual value provided by the platform and only
-  /// indicates a value was provided that we don't currently support.
+  /// indicates a value was provided that isn't currently supported.
   unknown,
 }
 
@@ -526,6 +533,9 @@ class NSErrorData {
   String localizedDescription;
 
   /// The value of the inherited property `NSObject.description`.
+  ///
+  /// The name `description` cannot be used because it is an inherited property
+  /// that is `readonly` and can't be set.
   String objectDescription;
 
   Object encode() {